path-uri: decouple native path parsing
